chore: refactor runtimes.ts
