+++ src/lib/wizards/functions/steps/functionConfiguration.svelte
@@ -7,7 +7,7 @@
     import { onMount } from 'svelte';
     import { sdk } from '$lib/stores/sdk';
     import { choices, createFunction, installation, repository } from '../store';
-    import { runtimesList } from '$routes/(console)/project-[project]/functions/store';
+    import { runtimesList } from '$lib/stores/runtimes';
 
     let showCustomId = false;
